:recycle: refactor : Base Entity 분리
