+++ com.rush.logistic.client.slack/src/main/java/com/rush/logistic/client/slack/domain/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.rush.logistic.client.domain.global;
+package com.rush.logistic.client.slack.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
